[CHORE]/#31 폴더 모듈 생성
